IE backdrop fixes
When backdrop was used on IE, a white background would appear at the bottom of the page. I was able to reproduce this issue in IE8, IE10 and IE11.

In IE8 the backdrop background color was black without any opacity.

These fixes were tested in IE8, IE10, IE11 and Chrome, everything seemed to be functioning properly in all four browsers.