Added skipOrphan option
By default skipOrphan is true, and the behavior of the boostrap-tour is not changed.
However if skipOrphan is false, then when the step element is not found, the tour will do nothing.
No step will be shown until the tour is reinitialized and step element is found.

Use case:
We want to go to step i + 1 after visitor click on the button, but we do not know where the button will take us.
What we know is that on that location will be the step element. Why we do not know where the button will take us?
Link on button will be to the /create/item but the server will redirect us somewhere to the /item/XY

So unless we do not know what XY will be, we can not user redirect option + regex path and therefore
boostrap-tour will skip all orphaned steps. And because during the click on the button there is no elements from other
steps, it will take as to the last step and will end the tour.

Solution to this is skipOrphan = false. On the button click the  boostrap-tour will do nothing, but because we know
that it will redirect us somewhere where the step element will be present, boostrap-tour will find the step element
there and properly show the step.